django_app/requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astunparse==1.6.3
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
Cython==3.0.12
Django==5.2.1
djangorestframework==3.16.0
fastapi==0.115.12
Flask==3.1.0
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
keras==3.9.2
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.9
numpy==1.26.0
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
patsy==1.0.1
pmdarima==2.0.4
protobuf==5.29.4
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.2
setuptools==80.3.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
statsmodels==0.14.4
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
